requirements.txt
azure-functions==1.16.0
azure-storage-blob==12.19.0

torch==2.2.1
torchvision==0.17.1
ultralytics==8.1.25

opencv-python==4.9.0.80
numpy==1.24.4
scikit-image==0.22.0
Pillow==9.5.0

supervision==0.18.0

requests==2.31.0
